UndoTree command with checkout

+++ Draw.c
@@ -13,7 +13,7 @@ void Draw_Start(char* description) {
 	}
 }
 
-void Draw_Block(short x, short y, short z, BlockID block) {
+void Draw_Block(int x, int y, int z, BlockID block) {
 	if (!IsInWorldBoundaries(x, y, z)) {
 		return;
 	}
